Performance rewrite
I don't know what I'm doing, but anecdotally I'm seeing the benchmark
consistently complete in about 70% of the time it used to take. My
approach was:

 - Force the use of a BufReader (not included in the percentage above,
   but it made some other optimisations possible)
 - Inline some particularily hot functions
 - Allocate just one big Vec up front, and use it for all decrunching,
   in order to:
   - Decrease the amount of dynamic Vec allocations
   - Decrease the amount of data being moved/copied around
 - Write an allocation-less header parser (a marginal improvement)

I'm not 100% positive that there isn't a valid use case for this crate
where you don't want a BufReader, but for my use case I definitely want
one. Reading megabytes of data one byte at a time is not a good idea.