Return LoadError::Eof when appropriate
Before, an unexpected EOF in a non-run in the new decrunch resulted in a
LoadError::FileFormat.
The #[cold] function is essential. Without it, the benchmark takes more
than twice as long. Constructing the IoError is apparently not something
we want to be doing in our critical loop. Though I'm considering
removing the inner error from the public API so I don't have to
construct a "fake" IoError at all. It doesn't seem to ever communicate
anything more than the fact that the file ended.